+++ spec/features/metadata_display_spec.rb
@@ -56,5 +56,18 @@ RSpec.feature 'Metadata display' do
         expect(page).to have_css('dd', text: 'Constituent note')
       end
     end
+
+    it 'can toggle all' do
+      click_link 'Expand all'
+      within '.mods_display_nested_related_items' do
+        expect(page).to have_css('dl', visible: true)
+        expect(page).to have_css('dt', text: /Note:/i)
+      end
+      click_link 'Collapse all'
+      within '.mods_display_nested_related_items' do
+        expect(page).to have_css('dl', visible: false)
+        expect(page).to have_css('li a', text: 'Constituent Title')
+      end
+    end
   end
 end